backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.35
python-multipart==0.0.12
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
stripe==10.12.0
python-dotenv==1.0.1
aiosmtplib==3.0.1
email-validator==2.2.0
pydantic[email]==2.9.2
